+++ packages/fighting-design/breadcrumb-item/src/interface.d.ts
@@ -1,0 +1,1 @@
+export type { BreadcrumbItemPropsType } from './props'