+++ packages/fighting-design/breadcrumb-item/src/props.ts
@@ -1,0 +1,32 @@
+import type {
+  ExtractPropTypes,
+  VNode,
+  Component,
+  PropType
+} from 'vue'
+
+export const Props = {
+  /**
+   * @zh 自定义文字颜色
+   */
+  color: {
+    type: String,
+    default: (): null => null
+  },
+  /**
+   * @zh 自定义分隔符颜色
+   */
+  iconColor: {
+    type: String,
+    default: (): null => null
+  },
+  /**
+   * @zh 自定义分隔符
+   */
+  separator: {
+    type: Object as PropType<VNode | Component>,
+    default: (): null => null
+  }
+}
+
+export type BreadcrumbItemPropsType = ExtractPropTypes<typeof Props>